myapp/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.0
djangorestframework==3.14.0
future==0.18.3
idna==3.6
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.0
urlsresolver==1.2.0
